Use integers for scores for consistancy Store scores and estimations in node

+++ Tree.cpp
@@ -34,6 +34,6 @@ Tree::Node Tree::findNode(Vec2<> pos) {
 }
 
 // Manhattan Distance
-float SearchAlgorithm::heuristic(Tree::Node start, Tree::Node end) {
+int SearchAlgorithm::heuristic(Tree::Node start, Tree::Node end) {
 	return (float)abs(start->pos.x - end->pos.x) + (float)abs(start->pos.y - end->pos.y);
 }
